[3.3.7.6] * Meer foutoplosingen en verbeteringen.

+++ Rpm/AutoUpdater/Forms/DownloadUpdateDialog.cs
@@ -16,11 +16,11 @@ namespace AutoUpdaterDotNET
     {
         private readonly UpdateInfoEventArgs _args;
 
+        private DateTime _startedAt;
+
         private string _tempFile;
 
         private MyWebClient _webClient;
-
-        private DateTime _startedAt;
 
         public DownloadUpdateDialog(UpdateInfoEventArgs args)
         {
@@ -28,10 +28,7 @@ namespace AutoUpdaterDotNET
 
             _args = args;
 
-            if (AutoUpdater.Mandatory && AutoUpdater.UpdateMode == Mode.ForcedDownload)
-            {
-                ControlBox = false;
-            }
+            if (AutoUpdater.Mandatory && AutoUpdater.UpdateMode == Mode.ForcedDownload) ControlBox = false;
         }
 
         private void DownloadUpdateDialogLoad(object sender, EventArgs e)
@@ -47,10 +44,7 @@ namespace AutoUpdaterDotNET
             else
             {
                 _tempFile = Path.Combine(AutoUpdater.DownloadPath, $"{Guid.NewGuid().ToString()}.tmp");
-                if (!Directory.Exists(AutoUpdater.DownloadPath))
-                {
-                    Directory.CreateDirectory(AutoUpdater.DownloadPath);
-                }
+                if (!Directory.Exists(AutoUpdater.DownloadPath)) Directory.CreateDirectory(AutoUpdater.DownloadPath);
             }
 
             _webClient.DownloadProgressChanged += OnDownloadProgressChanged;
@@ -62,14 +56,14 @@ namespace AutoUpdaterDotNET
 
         private void OnDownloadProgressChanged(object sender, DownloadProgressChangedEventArgs e)
         {
-            if (_startedAt == default(DateTime))
+            if (_startedAt == default)
             {
                 _startedAt = DateTime.Now;
             }
             else
             {
                 var timeSpan = DateTime.Now - _startedAt;
-                long totalSeconds = (long) timeSpan.TotalSeconds;
+                var totalSeconds = (long) timeSpan.TotalSeconds;
                 if (totalSeconds > 0)
                 {
                     var bytesPerSecond = e.BytesReceived / totalSeconds;
@@ -84,28 +78,17 @@ namespace AutoUpdaterDotNET
 
         private void WebClientOnDownloadFileCompleted(object sender, AsyncCompletedEventArgs asyncCompletedEventArgs)
         {
-            if (asyncCompletedEventArgs.Cancelled)
-            {
-                return;
-            }
+            if (asyncCompletedEventArgs.Cancelled) return;
 
             try
             {
-                if (asyncCompletedEventArgs.Error != null)
-                {
-                    throw asyncCompletedEventArgs.Error;
-                }
+                if (asyncCompletedEventArgs.Error != null) throw asyncCompletedEventArgs.Error;
 
-                if (_args.CheckSum != null)
-                {
-                    CompareChecksum(_tempFile, _args.CheckSum);
-                }
+                if (_args.CheckSum != null) CompareChecksum(_tempFile, _args.CheckSum);
 
                 ContentDisposition contentDisposition = null;
                 if (_webClient.ResponseHeaders["Content-Disposition"] != null)
-                {
                     contentDisposition = new ContentDisposition(_webClient.ResponseHeaders["Content-Disposition"]);
-                }
 
                 var fileName = string.IsNullOrEmpty(contentDisposition?.FileName)
                     ? Path.GetFileName(_webClient.ResponseUri.LocalPath)
@@ -116,19 +99,14 @@ namespace AutoUpdaterDotNET
                         string.IsNullOrEmpty(AutoUpdater.DownloadPath) ? Path.GetTempPath() : AutoUpdater.DownloadPath,
                         fileName);
 
-                if (File.Exists(tempPath))
-                {
-                    File.Delete(tempPath);
-                }
+                if (File.Exists(tempPath)) File.Delete(tempPath);
 
                 File.Move(_tempFile, tempPath);
 
                 string installerArgs = null;
                 if (!string.IsNullOrEmpty(_args.InstallerArgs))
-                {
                     installerArgs = _args.InstallerArgs.Replace("%path%",
                         Path.GetDirectoryName(Process.GetCurrentProcess().MainModule.FileName));
-                }
 
                 var processStartInfo = new ProcessStartInfo
                 {
@@ -140,28 +118,23 @@ namespace AutoUpdaterDotNET
                 var extension = Path.GetExtension(tempPath);
                 if (extension.Equals(".zip", StringComparison.OrdinalIgnoreCase))
                 {
-                    string installerPath = Path.Combine(Path.GetDirectoryName(tempPath), "ZipExtractor.exe");
+                    var installerPath = Path.Combine(Path.GetDirectoryName(tempPath), "ZipExtractor.exe");
 
                     File.WriteAllBytes(installerPath, Resources.ZipExtractor);
 
-                    string executablePath = Process.GetCurrentProcess().MainModule.FileName;
-                    string extractionPath = Path.GetDirectoryName(executablePath);
+                    var executablePath = Process.GetCurrentProcess().MainModule.FileName;
+                    var extractionPath = Path.GetDirectoryName(executablePath);
 
                     if (!string.IsNullOrEmpty(AutoUpdater.InstallationPath) &&
                         Directory.Exists(AutoUpdater.InstallationPath))
-                    {
                         extractionPath = AutoUpdater.InstallationPath;
-                    }
 
-                    StringBuilder arguments =
+                    var arguments =
                         new StringBuilder($"\"{tempPath}\" \"{extractionPath}\" \"{executablePath}\"");
-                    string[] args = Environment.GetCommandLineArgs();
-                    for (int i = 1; i < args.Length; i++)
+                    var args = Environment.GetCommandLineArgs();
+                    for (var i = 1; i < args.Length; i++)
                     {
-                        if (i.Equals(1))
-                        {
-                            arguments.Append(" \"");
-                        }
+                        if (i.Equals(1)) arguments.Append(" \"");
 
                         arguments.Append(args[i]);
                         arguments.Append(i.Equals(args.Length - 1) ? "\"" : " ");
@@ -181,16 +154,10 @@ namespace AutoUpdaterDotNET
                         FileName = "msiexec",
                         Arguments = $"/i \"{tempPath}\""
                     };
-                    if (!string.IsNullOrEmpty(installerArgs))
-                    {
-                        processStartInfo.Arguments += " " + installerArgs;
-                    }
+                    if (!string.IsNullOrEmpty(installerArgs)) processStartInfo.Arguments += " " + installerArgs;
                 }
 
-                if (AutoUpdater.RunUpdateAsAdmin)
-                {
-                    processStartInfo.Verb = "runas";
-                }
+                if (AutoUpdater.RunUpdateAsAdmin) processStartInfo.Verb = "runas";
 
                 try
                 {
@@ -199,13 +166,9 @@ namespace AutoUpdaterDotNET
                 catch (Win32Exception exception)
                 {
                     if (exception.NativeErrorCode == 1223)
-                    {
                         _webClient = null;
-                    }
                     else
-                    {
                         throw;
-                    }
                 }
             }
             catch (Exception e)
@@ -226,17 +189,17 @@ namespace AutoUpdaterDotNET
             string[] suf = {"B", "KB", "MB", "GB", "TB", "PB", "EB"};
             if (byteCount == 0)
                 return "0" + suf[0];
-            long bytes = Math.Abs(byteCount);
-            int place = Convert.ToInt32(Math.Floor(Math.Log(bytes, 1024)));
-            double num = Math.Round(bytes / Math.Pow(1024, place), 1);
+            var bytes = Math.Abs(byteCount);
+            var place = Convert.ToInt32(Math.Floor(Math.Log(bytes, 1024)));
+            var num = Math.Round(bytes / Math.Pow(1024, place), 1);
             return $"{(Math.Sign(byteCount) * num).ToString(CultureInfo.InvariantCulture)} {suf[place]}";
         }
 
         private static void CompareChecksum(string fileName, CheckSum checksum)
         {
             using (var hashAlgorithm =
-                HashAlgorithm.Create(
-                    string.IsNullOrEmpty(checksum.HashingAlgorithm) ? "MD5" : checksum.HashingAlgorithm))
+                   HashAlgorithm.Create(
+                       string.IsNullOrEmpty(checksum.HashingAlgorithm) ? "MD5" : checksum.HashingAlgorithm))
             {
                 using (var stream = File.OpenRead(fileName))
                 {
@@ -258,13 +221,12 @@ namespace AutoUpdaterDotNET
         private void DownloadUpdateDialog_FormClosing(object sender, FormClosingEventArgs e)
         {
             if (AutoUpdater.Mandatory && AutoUpdater.UpdateMode == Mode.ForcedDownload)
-            {
                 if (e.CloseReason == CloseReason.UserClosing)
                 {
                     e.Cancel = true;
                     return;
                 }
-            }
+
             if (_webClient is {IsBusy: true})
             {
                 _webClient.CancelAsync();
